Experiments with unboud callbacks.
I've changed the semantics of callback functions to make signal handlers
receive the same arguments as the "operation" function, plus additional
ones like "result", "progress" etc.

This allows binding methods as callbacks, but impacts callback signature
used with regular functions.

Not sure if this is good or bad, it might be that callbacks for free
functions should know what the arguments were... up for discussion I
guess.